+++ lib/example/controllers/user.controller.d.ts
@@ -1,6 +1,5 @@
 import { Promise } from "es6-promise";
-import { ApiController } from "../../easy-rest/controller/controller";
-import { IActionResult } from "../../easy-rest/controller/action-result/action-result";
+import { ApiController, IActionResult } from "../../index";
 import { Book } from "./book.controller";
 export declare class UserController extends ApiController {
     static books: Book[];