+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+<a name="0.2.2"></a>
+# [0.2.2](https://github.com/Odrin/express-easy-rest/compare/0.2.0...0.2.2) (2016-08-22)
+
+### Code Refactoring
+* fixed authorize decorator naming (authorize -> Authorize)
+* some fixes and refactoring in examples
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/Odrin/express-easy-rest/compare/0.1.0...0.2.0) (2016-08-19)